Extract some common methods into a module
