+++ lib/patronus_fati/data_models/common_state.rb
@@ -1,0 +1,34 @@
+module PatronusFati
+  module DataModels
+    module CommonState
+      def active?
+        presence.visible_since?(self.class.current_expiration_threshold)
+      end
+
+      def data_dirty?
+        sync_flag?(:dirtyAttributes) || sync_flag?(:dirtyChildren)
+      end
+
+      def dirty?
+        new? || data_dirty? || status_dirty?
+      end
+
+      def new?
+        sync_status == SYNC_FLAGS[:unsynced]
+      end
+
+      def set_sync_flag(flag)
+        self.sync_status |= SYNC_FLAGS[flag]
+      end
+
+      def status_dirty?
+        sync_flag?(:syncedOnline) && !active? ||
+          sync_flag?(:syncedOffline) && active?
+      end
+
+      def sync_flag?(flag)
+        (sync_status & SYNC_FLAGS[flag]) > 0
+      end
+    end
+  end
+end
